PCI: altera: set maximum supported TLP data payload size for Agilex 5
The Max Payload Size (MPS) in the Device Control register of the PCIe
root port defaults to 128 bytes. Increase the payload size to the Max
Payload Size Supported (MPSS) from the Device Capabilities register,
e.g., 256 bytes to reduce transaction overhead and increase bandwidth.

The PCIe subsystem negotiates the MPS of root port and endpoint based
on the MPS stored in the Device Control register of the root port. If
the MPS of the endpoint differs from that of the root port, the MPS
of the endpoint is changed to match that of the root port. In all
cases, the negotiated MPS is limited by the maximum of the MPSS
from the Device Capabilities registers of root port and end point.

Upstream-Status: Pending [Will be submitted soon]
Signed-off-by: Mahesh R Vaidya <mahesh.r.vaidya@intel.com>